refactor: Move meta components into their own module
